Grand value division: value::partial_eq
