+++ src/value/mod.rs
@@ -913,79 +913,7 @@ impl Default for Value {
 }
 
 mod index;
-
-impl PartialEq<str> for Value {
-    fn eq(&self, other: &str) -> bool {
-        self.as_str().map_or(false, |s| s == other)
-    }
-}
-
-impl<'a> PartialEq<&'a str> for Value {
-    fn eq(&self, other: &&str) -> bool {
-        self.as_str().map_or(false, |s| s == *other)
-    }
-}
-
-impl PartialEq<Value> for str {
-    fn eq(&self, other: &Value) -> bool {
-        other.as_str().map_or(false, |s| s == self)
-    }
-}
-
-impl<'a> PartialEq<Value> for &'a str {
-    fn eq(&self, other: &Value) -> bool {
-        other.as_str().map_or(false, |s| s == *self)
-    }
-}
-
-impl PartialEq<String> for Value {
-    fn eq(&self, other: &String) -> bool {
-        self.as_str().map_or(false, |s| s == other)
-    }
-}
-
-
-impl PartialEq<Value> for String {
-    fn eq(&self, other: &Value) -> bool {
-        other.as_str().map_or(false, |s| s == self)
-    }
-}
-
-macro_rules! partialeq_numeric {
-    ($([$($ty:ty)*], $conversion:ident, $base:ty)*) => {
-        $($(
-            impl PartialEq<$ty> for Value {
-                fn eq(&self, other: &$ty) -> bool {
-                    self.$conversion().map_or(false, |i| i == (*other as $base))
-                }
-            }
-
-            impl PartialEq<Value> for $ty {
-                fn eq(&self, other: &Value) -> bool {
-                    other.$conversion().map_or(false, |i| i == (*self as $base))
-                }
-            }
-
-            impl<'a> PartialEq<$ty> for &'a Value {
-                fn eq(&self, other: &$ty) -> bool {
-                    self.$conversion().map_or(false, |i| i == (*other as $base))
-                }
-            }
-
-            impl<'a> PartialEq<$ty> for &'a mut Value {
-                fn eq(&self, other: &$ty) -> bool {
-                    self.$conversion().map_or(false, |i| i == (*other as $base))
-                }
-            }
-        )*)*
-    }
-}
-
-partialeq_numeric! {
-    [i8 i16 i32 i64 isize], as_i64, i64
-    [u8 u16 u32 u64 usize], as_u64, u64
-    [f32 f64], as_f64, f64
-}
+mod partial_eq;
 
 macro_rules! from_integer {
     ($($ty:ident)*) => {
